requirements.txt
pillow==11.0.0
pygame==2.6.1
pygame-ce==2.5.2
pygame-pgu==0.21
pygame_gui==0.6.12
python-i18n==0.3.9
tinydb==4.8.2
